Add send_volatile for WebSocket connections
I have a project which uses realtime data to some extent, such as mouse
positions. For that type of data it’s important to me that the peer has
the most up-to-date data possible, and it’s unimportant, and unwanted
that the data gets queued and delivered in a guaranteed manner.

Therefore I added send_volatile(), which only works for WebSockets. It
checks if the WebSocket buffer is empty before sending. For other
transports it just sends as normal.